Refactor GUI tests into multiple modules
